GEODE-6534: Extract ExpiryRegionEntry from LocalRegion
